feat(backstage-plugin): Improve ui layout
Change the layout to using a simpler flexbox styling with inline styling so it's easier to follow than a bunch of embedded grids.

Add responsive sizing to the graphs.

Stick aggregation controls to top of the screen when scrolling.

Still TODO: Add controls for selecting which data points to show.